power_suspend: Whitelist touchpad early wakeup problem on Link
Link suffers from occasional spurious wakeups due to a touchpad
interrupt (GPIO 12). This was determined to be mostly harmless and to
tricky to fix, so let's just whitelist our suspend tests to retry during
that condition.

This will also improve log aggregation for EarlyWakeup errors by
removing the timestamp and adding the wakeup reason to the autotest
failure reason string.

BUG=chromium:220014
TEST=power_Resume
	1. Run normally on Link and something else, should succeed
	2. Run on something else and wake it early, should fail
	3. Run on Link and hit keyboard, should fail
	4. Run on Link and touch touchpad, should retry

Change-Id: I3a28da81db9e850603d606f32e23973f316a5799
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/56538