firmware_TouchMTB: add an option not to display the html result
When we modify what the content or format that is output by the test
result log, we need to replay the gestures to derive the log of the
new format. In this case, we do not want to display the html test
result on the chrome browser which takes significant time by
telemetry to logout and login to the chrome with a test account.

In this patch, we add a command option not to display the html
test result.

BUG=chromium:241092
TEST=Run the following commands:
$ cd $ cd /usr/local/autotest/tests/firmware_TouchMTB
$ mkdir -p /var/tmp/touch_firmware_test
$ cp -r tests/logs/lumpy/20130506_030025-fw_11.27-robot_sim/ \
  /var/tmp/touch_firmware_test
$ OPTIONS="-m robot_sim --skip_html -i 2 --replay
           20130506_032659-fw_11.23-robot_sim" python main.py
Observe that the test should not invoke telemetry at the end.

Change-Id: I55280c552769c94e73d3f5fcfafdeffbd9a5e5d3
Reviewed-on: https://gerrit.chromium.org/gerrit/56728
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>