firmware_TouchMTB: add an option to specify the log root in summary.sh
In the summary.sh, it assumes that the logs reside in /var/tmp. It
is desirable to be able to specify the log root instead of the fixed
one. This is useful when we would like to derive a summary from
the logs that are replayed and re-generated in other directory.

In this patch, we add an option in summary.sh that allows a user
to specify the log root.

BUG=chromium:241092
TEST=Run the following commands
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ mkdir -p /tmp/touch_firmware_test
$ cp -r tests/logs/lumpy/20130506_030025-fw_11.27-robot_sim/ \
  /tmp/touch_firmware_test
 # Use /tmp instead of /var/tmp as the log root.
$ ./summary.sh -l /tmp
Observe that the summary is stored in /tmp/summary.

Change-Id: I4d6f60fe3392b2e0c835d9ebbaa9cacab5565de3
Reviewed-on: https://gerrit.chromium.org/gerrit/56729
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>