firmware_TouchMTB: add scripts to replay and copy logs
Whenever the contents or formats of the logs are changed, we need to
replay and re-generate the logs. In this patch, we automate the flow
of replaying logs on the remote machine, and scp the logs to the host.

When a user executes tools/host_replay.sh on a host, it ssh to the
specified remote machine to execute tools/machine_replay.sh to replay
the logs, removes the old logs on the host, and then scp the new logs
from the machine to the host.

BUG=chromium:241092
TEST=Run the following commands in chroot
(cr) $  tools/host_replay.sh -m $MACHINE_IP
Observe that the specified machine replays and generates new logs.
And then those logs are scp-ed to the host.

Change-Id: I5f97f3f533feef0839a6cd4bd8e47564195b2c12
Reviewed-on: https://gerrit.chromium.org/gerrit/56730
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>