[autotest] Add provisioning flag to autoserv.
autoserv now accepts a --provision flag that accepts a comma seperated
list of labels that it should provision the targeted machine(s) (via -m)
to.  This list gets fed directly into the control segment, which handles
all of the processing of it and provisioning work.

It should also be loudly noted that since we run the provision control
segment as a normal test instead of directly executing the contents, one
cannot specify the -c flag, or else it will try to run the provision
control segment as a client-side test.

This is just some plumbing.  It's pretty uneventful.

BUG=chromium:249437
TEST=The following:
1) Ran autoserv --provision cros-version:lumpy-release/R28-3993.0.0 and
the machine was provisioned correctly.
2) Ran autoserv -c --provision and received an error.

Change-Id: I03582a5210d182f171ef725c4e49cc83c1baf793
Reviewed-on: https://gerrit.chromium.org/gerrit/58384
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>