FAFT: Should not call a check using RPC if DUT may fail to start faft_client
We saw a failure that DUT is online but fails to start faft_client. In the
reset rountine, the cold reboot doesn't help. It then goes to recovery
boot rountine. In the recovery boot rountine, we should not call the check
which uses RPC because the faft_client may not start. A BadStatusLine
exception is raised during the RPC call.

Directly doing recovery boot and not check fixes this problem.

BUG=chromium:239034
TEST=None, trivial change, hard to repro the original failure case.

Change-Id: I4773deb6148d310478541ad3ba1dfd3d3241b44b
Reviewed-on: https://gerrit.chromium.org/gerrit/57103
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>