logging_UserCrash: max_enqueued_crashes: poll for syslog messages
There is a small race condition where we trigger a crash and then search
for the log message, but syslog hasn't had a chance to write the data.
It's pretty rare, but it's easy to fix, so do so.  Use the handy helper
func to poll the log files for things.

BUG=chromium:201975
TEST=logging_UserCrash test passes
TEST=manually mucking up error message shows the code polls before timing out & failing

Change-Id: I249e18b0f2df10bb701a7599ba993bbaae5f55b6
Reviewed-on: https://gerrit.chromium.org/gerrit/57114
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>