[autotest] Bug reporter uses apiary backend to create new issues.
This cl executes the second part of the bug reporters transition to the
new apiary api. To create new issues we use phapi_lib (the apiary backend)
and to comment on existing issues we use gdata_lib. In time we will
transition completely to using phapi_lib.

TEST=Created new issues with Restrict-View-Google then made sure we can
     dedupe against them.
BUG=chromium:241510

Change-Id: I817b609091eabdf96fd056b22cd58ca73b01f266
Reviewed-on: https://gerrit.chromium.org/gerrit/57547
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>