Add L2TP/IPSec client test
This is a client-only test which verifies function of the L2TP/IPSec
client.  Since the server it is testing against is itself, this test
is best suited for verifying that the connection manager and VPN
manager are correctly passing parameters to the underlying VPN client
to have it properly connect.

BUG=chromium:243147
TEST=This is a test

Change-Id: Ifc1280b085ab91379556f69d20fa48a4da2bd0ab
Reviewed-on: https://gerrit.chromium.org/gerrit/57270
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>