FAFT: Stop update-engine when running FAFT
FAFT tests the behaviors of firmware and kernel. We don't want the
update-engine touches the firmware and kernel while FAFT is running.
So this CL stops update-engine on every boot to make the test simple.

BUG=chrome-os-partner:19727
TEST=Manal
Run a FAFT test and check the following message showed on every boot:
02:23:32 INFO | Stopping update-engine...

Change-Id: I73ed219276b185d28cb71ea833c3c4f2d0707787
Reviewed-on: https://gerrit.chromium.org/gerrit/57308
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>