[autotest] Add a cros-version provisioning test.
This adds a test similar to the autoupdate test that can reimage a
machine to a requested build, except it adheres to what the expected API
is for a provisioning test.

BUG=chromium:249437
TEST=The following:
1) Ran via autoserv against a machine without a cros-version label,
and the machine was correctly reimaged to the requested version.
2) Ran via autoserv against a machine with a cros-version label, but of
a different image, and the machine was correctly reimaged to the
requested version.
3) Ran via autoserv against a machine with the correct cros-verion
label, and the test correctly returned with TEST_NA.
4) Ran via autoserv with malformed --args (ie.
value=lumpy-release/R28-3993.0.0.0, value=lum~py-release/R28-3993.0.0)
and saw a devserver error.

Change-Id: I74afc05eb8b6302a27ec37e8c2cca7b21c0471e1
Reviewed-on: https://gerrit.chromium.org/gerrit/58382
Commit-Queue: Alex Miller <milleral@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>