[autotest] Make board always be the metahost.
When the scheduler goes to pull the list of hosts to consider for a HQE
to run against, it pulls all hosts with the label specified by metahost.
Currently, we use the cros-version label, if available, and pool
otherwise.  In preperation for preemptive scheduling, we need the
scheduler to consider all hosts within any given board, since it can
(plus or minus some DEPENDENCIES) choose any of them to provision for
the job.

It's worth noting that this CL might also speed up the scheduler tick,
as we're going to be considering ~30 hosts per job, instead of ~100.
(Average number of hosts per board vs average number of hosts per pool.)

There's also some rework hidden in here to add |board| as a required
argument to Suite.

BUG=chromium:249437
TEST=local run_suite, unittest

Change-Id: I4afd8ef958b96c9fb9bc4a7f838a360d508683d6
Reviewed-on: https://gerrit.chromium.org/gerrit/58386
Commit-Queue: Alex Miller <milleral@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>