[autotest] Do not consider provisionable labels when scheduling AND do not schedule tryjobs for cros-version.
We now have the ability to schedule provisioning tasks, but we don't as
the labels on a job always perfectly match the labels on a host.  We can
now disregard provisionable labels when scheduling, as provisioning will
take care of that for us.  This CL will have some interesting effects,
as it's not very intelligent as to what host it chooses to provision.

This CL also has turning cros-version tryjobs into a no-op folded into
it.  It should be its own CL, but the two would have to get landed (and
possibly reverted) together, so thus it seemed far easier two merge the
two so that there aren't any mixups.

BUG=chromium:249437, chromium:250586
DEPLOY=scheduler
TEST=Jobs with provisionable labels are scheduled against hosts without
those labels.  Dynamic suites no longer schedules tryjobs for reimaging
a machine, and the suite still runs and completes successfully.

Change-Id: Ia1793c0d9cdcd9f37d1392dbcbe5af12550d9791
Reviewed-on: https://gerrit.chromium.org/gerrit/58387
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>