firmware_TouchMTB: add metrics for PhysicalClickValidator
  We add the metrics for PhysicalClickValidator to show the
percentage of finger IDs without a click. To achieve that,
we add a number of one-finger and two-finger event files with
physical clicks, update the result logs, add unit tests for
physical clicks, and update the unit tests for summary tests.

BUG=chromium:251193
TEST=Replay logs on a lumpy.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ tools/machine_replay.sh -b tests/logs/lumpy/ -s
And then run the summary script to generate a summary report.
$ ./summary.sh -d /tmp -m

Observe its metrics would be something like

PhysicalClickValidator
  fw_11.23: {'2f_click_hit_rate_%': (0.0, 0.0, 0.0)}
  fw_11.27: {'2f_click_hit_rate_%': (1.0, 1.0, 1.0)}

Note that we show the hit rates instead of the failure rates so
that the larger the numbers the better the results which are
consistent with all of the other numbers elsewhere.

Also all of the unit tests should pass.
$ python tests/run_all_unittests.py

Change-Id: I972de131f02066c5a639f1f4b42b517d3fed6ebd
Reviewed-on: https://gerrit.chromium.org/gerrit/59002
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>