Verify that the TPM driver contains the "gentle shutdown" patch.
The "gentle shutdown" patch to the TPM driver fixes (or reduces)
a serious problem that causes a certain TPM model to go into
self-defense mode for a long time.  Originally this test made
sure (among other things) that the driver was installed.  Now
it also checks that the driver running in the latest boot
session contains the "gentle shutdown" patch.

BUG=chromium:234976
TEST=tested for both success and failure

Change-Id: I5a47c3d0a993b83b3f9970753acb586d021d5d29
Reviewed-on: https://chromium-review.googlesource.com/58920
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>