firmware_TouchMTB: metrics for RangeValidator
In this patch, we add the number of mm that cannot be reached
for each of the four edges.

BUG=chromium:251604
TEST=Replay logs on a lumpy.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ tools/machine_replay.sh -b tests/logs/lumpy/ -s
And then run the summary script to generate a summary report.
$ ./summary.sh -d /tmp -m

Observe its metrics would be like:
  RangeValidator
    fw_11.23: {'short_of_range_left_mm': (0.083, 0.083, 0.083)}
    fw_11.27: {'short_of_range_left_mm': (0.0, 0.0, 0.0)}

All of the unit tests should pass.
$ python tests/run_all_unittests.py

Change-Id: I315faa1eb72d84ee522f56ae13334a559b8b4ff8
Reviewed-on: https://gerrit.chromium.org/gerrit/59217
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>