firmware_TouchMTB: add metrics for Stationary Finger Validator
In this patch, we add the metric for StationaryFingerValidator:
- the maximum distance between any two points of the stationary finger in mm

The result logs are updated due to the change of the validator and
its criteria. Some unit tests are also added/updated accordingly.

BUG=chromium:252669
TEST=Replay logs on a lumpy.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ tools/machine_replay.sh -b tests/logs/lumpy/ -s
And then run the summary script to generate a summary report.
$ ./summary.sh -d /tmp -m

Observe the metrics would look like:
ReportRateValidator
  fw_11.23: {
    'max_distance_mm': (1.3453, 1.3453, 1.3453)
  }
  fw_11.27: {
    'max_distance_mm': (1.7159, 1.7159, 1.7159)}
  }

All of the unit tests should pass.
$ python tests/run_all_unittests.py

Change-Id: Ia010ba136434348762ba2ae3351e5ac929dc28ec
Reviewed-on: https://gerrit.chromium.org/gerrit/59723
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>