[autotest] remove hard pycurl dependency from rpm_control_system
pycurl is not installed in the chroot, and installing it into the chroot
is not straightforward. This was causing rpm_control_system unit tests
to fail inside the chroot.

Fortunately, we aren't apparently even using pycurl in our
infrastructure (see dli_urllib, which creates a subclass of
dli.powerswitch that uses urllib instead of pycurl). So, by ignoring the
ImportError failure in importing pycurl, the unittests can run happily
even in the chroot.

BUG=chromium:221254
TEST=rpm_controller_unittest and rpm_dispather_unittest both pass inside
the chroot

Change-Id: I1d55bb3b7a8bb099e0147b7769c51831b7df45e0
Reviewed-on: https://gerrit.chromium.org/gerrit/59604
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>