Suite Scheduler shouldn't send 'lab is down' emails.
The lab is only down if we manually set it to lab it down hence we
already know it's down. This makes the stream of suite scheduler emails
pointless.

Fixing by changing the log level of this message from Warn to Debug
as we only send emails for level Warn or Error.

BUG=chromium:244522
TEST=Unittests.

Change-Id: If2bf24935bbcd96870d3aff7b5131537e0e90312
Reviewed-on: https://gerrit.chromium.org/gerrit/59972
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>