[autotest] Skip setting a sigalarm if we can't set a handler for it.
Our retry decorator is currently designed to use signals to timeout
blocking calls, and throw a timeout exception when such signals hit.
The timeout exception is handled in the caller, which may choose to
move on ignoring it, or raise it as necessary. Unfortunately wsgi
silently intercepts any attempt to install a sigalarm handler to
protect apache. This means that when an rpc call is made through wsgi,
we will feel the full force of a sigalarm and the caller has no say
in how we handle the timeout. Moreover, if we were to hit a sigalarm
within a nested retry in an rpc call, the outer retry will continue
to attempt the failed call from scratch causing a vicious cycle.

This cl modifies the retry decorator to only set a sigalarm if
installation of the handler was successfull. If we fail to install
the handler we directly call the method decorated by the retry,
assuming it has an internal timeout (like a socket timeout). The cl
also refactors some code around the retry logic to return in time
for a deadline, by only passing the remaining time to the sigalarm
decorator.

Note that there are 3 timeouts in play here:
1. The outer frontend_wrapper retry timeout (which is enforced by a sigalarm) 
2. The inner devserer timeout which (which will be enforced by a socket timeout)
3. Devserver retry timeout (which should be enforced by a sigalarm when possible)

More generally, anything that can be called both from within the context of wsgi
and from outside wsgi, should only set a sigalarm in the latter case and skip
setting one in the former case.

BUG=chromium:246209
TEST=
a. Timeout 3 will fire and we get a sigalarm if max(1,2) > 3 if we bypass wsgi
b. Timeout 3 won't fire and we will wait for min(1,2) if we're going through wsgi
c. Timeout 1 gets decrementally smaller and we eventually breakout of the loop,
   if we keep hitting b with 1>2.
d. min(1, 2) fires.
e. ran retry_unittests

Change-Id: I8d7ed21b8b4989b91f7581f0916559c5d93faa80
Reviewed-on: https://gerrit.chromium.org/gerrit/60106
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>