RPM Infrastructure: Be able to disable emails.
Currently if something is wrong with the RPM infrastructure we can
get plagued with large amounts of emails in a short period of time.

Being able to disable the emails will be useful for when problems
occur or when debugging the live infrastructure can cause emails.

Users can now disable or enable the emails from their own checkout:
./rpm_client --disable_emails=12
./rpm_client --enable_emails

Part of this change includes making rpm_client an executable script,
which also makes it easier for users to call the RPM infrastructure
to make outlet state changes.

./rpm_client --machine='chromeos1-rack5-host1' --state='OFF'

BUG=chromium:238411
TEST=local setup on my machine. Ensured that email disabling and
enabling works as expected.

Change-Id: Id138fc419a865969fcacef83631ed1d383026129
Reviewed-on: https://gerrit.chromium.org/gerrit/59991
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>