firmware_TouchMTB: add pressure to MTB state machine
In order to support a firmware pressure test later, we add the
pressure attribute to the MTB state machine.

We also re-organize the related structures, including TidPacket and
TidData, a little bit so that
(1) the client functions of the state machine do not need to pack
    the packet data by themselves, and
(2) it becomes simpler to add more attributes (e.g.,
    ABS_MT_TOUCH_MAJOR for palm/thumb detection) later when needed.

BUG=chromium:221864
TEST=Run the unit tests on a lumpy.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ python tests/run_all_unittests.py
All of the unit tests should pass.

Change-Id: I0c2d62530df9d84e5ec59d9a82fa2147ef05b51e
Reviewed-on: https://gerrit.chromium.org/gerrit/60210
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>