Change update version checking when update_version is None
If a ChromiumOSUpdater instance is created with the
local_devserver=True option, the check_version_to_confirm_install()
method will always return False.  In the case of a local
devserver, we want to skip the version test, so this changes
the method to always return True for the local devserver case.

BUG=None
TEST=run_remote_tests with autoupdate/control.local

Change-Id: Ie0790e0e1498aaf8c60a2275829a4cbd35de454a
Reviewed-on: https://gerrit.chromium.org/gerrit/60151
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>