[autotest] fix 2 remaining unit tests that fail on builder
1) common_util_test has a test that assumes the USER environment
variable is set. However, this envvar is in fact NOT set in the builder
environment. The test has been modified to catch the resulting KeyError
and skip the test.

2) full_release_test fails on the builder due to an exception in
importing the dev_server module. For now I'm adding this test to the
SKIP_TEST list, and filing a bug to get that import logic sorted out.

BUG=chromium:221254
TEST=ran a tryjob, tests passed on builder.

Change-Id: Id0d3369e6529479235388e5dde8f78af9ea9d545
Reviewed-on: https://gerrit.chromium.org/gerrit/60163
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>