power_status: Add file path check for reading CPUFreq stats
The pre-stored time_in_state path maybe not exist when reading
CPUFreq stats if the CPU Hotplug/CPU Quiet funciton is enabled
and related CPUs are turned off. This will cause some tests are
failed if these tests need to read the CPUFreq stats.

This change checks whether the path is exist or not and avoid
to read non-exist file path.

BUG=None
TEST=Run power_LoadTest and test was passed on CPUQuiet enabled
system

Change-Id: I0c24ab77ed1a94246b4374927f9da8ef155bb586
Signed-off-by: Penny Chiu <pchiu@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/60334
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>