[autotest] test_that pull test results to subdirectory of /tmp
This CL causes test_that to pull logs to a unique-per-run subdirectory
of /tmp. In a future CL, this will allow test_that to parse the result
logs and summarize the success/failure of tests and suites.

BUG=chromium:243431
TEST=unit tests pass; Ran a suite with test_that, and saw that
/tmp/test_that_results_[unique string]/ was created, contained test
result logs, and had appropriate permissions to allow non-root user to
examine directory and results.

Change-Id: Id44d85946657b9d570cbf728b19018643512c9dd
Reviewed-on: https://gerrit.chromium.org/gerrit/60783
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>