Add OpenVPN authentication test with username/password
Add an OpenVPN test script that ensures that a username and
password supplied as arguments to the VPN service are used
to authenticate the client to a VPN server.  This verifies
the correct function of both the OpenVPN driver and the
management server within shill.  This will be useful for
regression testing with later versions of OpenVPN.

BUG=chromium:248539
TEST=Run this test

Change-Id: Ib987c1038daa43b86dbf07caef93c327e1bdf663
Reviewed-on: https://gerrit.chromium.org/gerrit/60705
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>