[autotest] Disable lab errors for aborts.
All aborts should have been converted to failures. We
still see some cases where aborts happen (eg: crbug.com/256791).
In these cases it's better to file a general bug instead of a
bug on the owner of the bug filer, so it gets the attention
of the sheriffs.

TEST=Confirmed that we still don't file general bugs for SERVER_JOB
     failures.
BUG=None

Change-Id: I6b50bf9f6537bab93fb4b54a94b8b40e498d323a
Reviewed-on: https://gerrit.chromium.org/gerrit/60802
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>