Add DHCP static-IP test
This test verifies that static IP parameters are applied to the
client's network interface if they are provided to the service.
It further ensures that during static IP negotiation although
a lease is negotiated, it is released back to the DHCP server.
When static IP configuration is removed another negotiation
should go back to using the server-provided IP address.

BUG=chromium:254230
TEST=This is a test

Change-Id: I2f739520203088495cb4667de8b94784e9362604
Reviewed-on: https://gerrit.chromium.org/gerrit/60815
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>