firmware_TouchMTB: printing the statistics of metrics
In this patch, we print the statistics of metrics by validators in
the summary report.

We also handle the metrics names in a more unified way. Before,
the same metric name strings, e.g., 'max error in x (mm)', were used
across different modules. Now with a new class MetricNameProps to
keep information about the metric names, the descriptions, and the
stat_func used to calculate the statistics of metrics, the metric
names could be derived consistently by expanding from a template, say
'max error in {} (mm)', with its variations, e.g., ['x', 'y'], to
derive the names: 'max error in x (mm)' and 'max error in y (mm)'.

We also remove the PhysicalClickValidator temporarily which incurs
divide-by-0 error. This validator is not specified in the spec v2.
However, it might be useful to add it back later to make sure that
a tap does not incur physical click.

The metrics in a summary report looks like:

Metrics statistics by validator : fw1.2  description
--------------------------------------------------------------------------------
  DrumrollValidator
    circle radius (mm)        :    1.37  Anything over 2mm is failure
  Linearity(BothEnds)Validator
    max error in x (mm)       :    9.34  The max err of all samples
    max error in y (mm)       :    6.33  The max err of all samples
    rms error in x (mm)       :    2.46  The mean of all rms means of all trials
    rms error in y (mm)       :    1.03  The mean of all rms means of all trials
  Linearity(Middle)Validator
    max error in x (mm)       :    4.91  The max err of all samples
    max error in y (mm)       :    6.30  The max err of all samples
    rms error in x (mm)       :    1.27  The mean of all rms means of all trials
    rms error in y (mm)       :    0.68  The mean of all rms means of all trials
  PhysicalClickValidator
    1f-click miss rate (%)    :    0.00  Should be close to 0 (0 is perfect)
    2f-click miss rate (%)    :    0.00  Should be close to 0 (0 is perfect)
    3f-click miss rate (%)    :    0.00  Should be close to 0 (0 is perfect)
    4f-click miss rate (%)    :    0.00  Should be close to 0 (0 is perfect)
    5f-click miss rate (%)    :    0.00  Should be close to 0 (0 is perfect)
  RangeValidator
    left edge not reached (mm):    0.08  Min unreachable distance
  ReportRateValidator
    average time interval (ms):   12.19  less than 1/60 sec is required
    intervals > 1/60 sec (%)  :    4.45  0% is required
    max time interval (ms)    :   57.96  less than 1/60 sec is required
  StationaryFingerValidator
    max distance (mm)         :    3.85  max distance of any two points from
                                         any run

BUG=chromium:235753
TEST=Run the command on a chromebook machine to generate a summary report.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
Replay the logs
$ tools/machine_replay.sh -b tests/logs/lumpy -s
Generate the summary report
$ ./summary.sh -m -d /tmp
Observe that the metrics are printed as above.

Also make sure all unit tests pass.
$ python tests/run_all_unittests.py

Change-Id: If7224354ef866cc3468752caa32feb1439b0b8cc
Reviewed-on: https://gerrit.chromium.org/gerrit/61883
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>