Fix ignore-status stanza for WiFi_Regulatory pings
The "ignore_status" parameter was added to the wrong ping in the
test.  Move it to the correct spot.  Also adjust the timeout value
for the ping command to match the expected behavior of the "ping"
command in the absence of a reply to the last packet.

BUG=chrome-os-partner:19953
TEST=Rerun network_WiFi_Regulatory on a host that drops all pings

Change-Id: Ibbb8c5ff714b402618b9c79c15bb69d550662e09
Reviewed-on: https://gerrit.chromium.org/gerrit/61216
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>